+++ backend/src/neuroagent/app/app_utils.py
@@ -22,9 +22,13 @@ from starlette.status import HTTP_401_UNAUTHORIZED
 from neuroagent.app.config import Settings
 from neuroagent.app.database.sql_schemas import Entity, Messages, Threads, utc_now
 from neuroagent.app.schemas import (
+    AnnotationMessageVercel,
+    AnnotationToolCallVercel,
     MessagesRead,
     MessagesReadVercel,
     PaginatedResponse,
+    RateLimitInfo,
+    ReasoningPartVercel,
     TextPartVercel,
     ToolCallPartVercel,
     ToolCallVercel,
@@ -300,11 +304,19 @@ def format_messages_vercel(
 ) -> PaginatedResponse[MessagesReadVercel]:
     """Format db messages to Vercel schema."""
     messages: list[MessagesReadVercel] = []
-    tool_call_buffer: list[dict[str, Any]] = []
+    parts: list[TextPartVercel | ToolCallPartVercel | ReasoningPartVercel] = []
+    annotations: list[AnnotationMessageVercel | AnnotationToolCallVercel] = []
 
     for msg in reversed(db_messages):
         if msg.entity in [Entity.USER, Entity.AI_MESSAGE]:
-            text_content = json.loads(msg.content).get("content")
+            content = json.loads(msg.content)
+            text_content = content.get("content")
+            reasoning_content = content.get("reasoning")
+
+            # Optional reasoning
+            if reasoning_content:
+                parts.append(ReasoningPartVercel(reasoning=reasoning_content))
+
             message_data = {
                 "id": msg.message_id,
                 "role": "user" if msg.entity == Entity.USER else "assistant",
@@ -313,58 +325,45 @@ def format_messages_vercel(
             }
             # add tool calls and reset buffer after attaching
             if msg.entity == Entity.AI_MESSAGE:
-                message_data["parts"] = [
-                    *[
-                        ToolCallPartVercel(toolInvocation=ToolCallVercel(**tool_call))
-                        for tool_call in tool_call_buffer
-                    ],
-                    TextPartVercel(text=text_content),
-                ]
-                message_data["annotations"] = [
-                    {"message_id": msg.message_id, "isComplete": msg.is_complete},
-                    *[
-                        {
-                            "toolCallId": tool_call["toolCallId"],
-                            "validated": tool_call["validated"],
-                            "isComplete": tool_call["is_complete"],
-                        }
-                        for tool_call in tool_call_buffer
-                    ],
-                ]
-                tool_call_buffer = []
-            # If we encounter a user message with a non empty buffer we have to add a dummy ai message.
-            elif tool_call_buffer:
-                last_tool_call = tool_call_buffer[-1]
-                messages.append(
-                    MessagesReadVercel(
-                        **{
-                            "id": uuid.uuid4().hex,
-                            "role": "assistant",
-                            "createdAt": last_tool_call["creation_date"],
-                            "content": "",
-                            "parts": [
-                                ToolCallPartVercel(
-                                    toolInvocation=ToolCallVercel(**tool_call)
-                                )
-                                for tool_call in tool_call_buffer
-                            ],
-                            "annotations": [
-                                {
-                                    "toolCallId": tool_call["toolCallId"],
-                                    "validated": tool_call["validated"],
-                                    "isComplete": tool_call["is_complete"],
-                                }
-                                for tool_call in tool_call_buffer
-                            ],
-                        }
+                if text_content:
+                    parts.append(TextPartVercel(text=text_content))
+
+                annotations.append(
+                    AnnotationMessageVercel(
+                        messageId=msg.message_id, isComplete=msg.is_complete
                     )
                 )
-                tool_call_buffer = []
 
+                message_data["parts"] = parts
+                message_data["annotations"] = annotations
+
+            # If we encounter a user message with a non empty buffer we have to add a dummy ai message.
+            elif parts:
+                messages.append(
+                    MessagesReadVercel(
+                        id=uuid.uuid4().hex,
+                        role="assistant",
+                        createdAt=msg.creation_date,
+                        content="",
+                        parts=parts,
+                        annotations=annotations,
+                    )
+                )
+
+            parts = []
+            annotations = []
             messages.append(MessagesReadVercel(**message_data))
 
         # Buffer tool calls until the next AI_MESSAGE
         elif msg.entity == Entity.AI_TOOL:
+            content = json.loads(msg.content)
+            text_content = content.get("content")
+            reasoning_content = content.get("reasoning")
+
+            # Add optional reasoning
+            if reasoning_content:
+                parts.append(ReasoningPartVercel(reasoning=reasoning_content))
+
             for tc in msg.tool_calls:
                 requires_validation = tool_hil_mapping.get(tc.name, False)
                 if tc.validated is True:
@@ -376,17 +375,23 @@ def format_messages_vercel(
                 else:
                     status = "pending"
 
-                tool_call_buffer.append(
-                    {
-                        "toolCallId": tc.tool_call_id,
-                        "toolName": tc.name,
-                        "args": json.loads(tc.arguments),
-                        "is_complete": msg.is_complete,
-                        "state": "call",
-                        # Needed for dummy messsages
-                        "validated": status,
-                        "creation_date": msg.creation_date,
-                    }
+                parts.append(TextPartVercel(text=text_content or ""))
+                parts.append(
+                    ToolCallPartVercel(
+                        toolInvocation=ToolCallVercel(
+                            toolCallId=tc.tool_call_id,
+                            toolName=tc.name,
+                            args=json.loads(tc.arguments),
+                            state="call",
+                        )
+                    )
+                )
+                annotations.append(
+                    AnnotationToolCallVercel(
+                        toolCallId=tc.tool_call_id,
+                        validated=status,  # type: ignore
+                        isComplete=msg.is_complete,
+                    )
                 )
 
         # Merge the actual tool result back into the buffered part
@@ -394,40 +399,39 @@ def format_messages_vercel(
             tool_call_id = json.loads(msg.content).get("tool_call_id")
             tool_call = next(
                 (
-                    item
-                    for item in tool_call_buffer
-                    if item["toolCallId"] == tool_call_id
+                    part.toolInvocation
+                    for part in parts
+                    if isinstance(part, ToolCallPartVercel)
+                    and part.toolInvocation.toolCallId == tool_call_id
+                ),
+                None,
+            )
+            annotation = next(
+                (
+                    annotation
+                    for annotation in annotations
+                    if isinstance(annotation, AnnotationToolCallVercel)
+                    and annotation.toolCallId == tool_call_id
                 ),
                 None,
             )
             if tool_call:
-                tool_call["result"] = json.loads(msg.content).get("content")
-                tool_call["state"] = "result"
-                tool_call["is_complete"] = msg.is_complete
+                tool_call.result = json.loads(msg.content).get("content")
+                tool_call.state = "result"
+
+            if annotation:
+                annotation.isComplete = msg.is_complete
 
     # If the tool call buffer is not empty, we need to add a dummy AI message.
-    if tool_call_buffer:
-        last_tool_call = tool_call_buffer[-1]
+    if parts:
         messages.append(
             MessagesReadVercel(
-                **{
-                    "id": uuid.uuid4().hex,
-                    "role": "assistant",
-                    "createdAt": last_tool_call["creation_date"],
-                    "content": "",
-                    "parts": [
-                        ToolCallPartVercel(toolInvocation=ToolCallVercel(**tool_call))
-                        for tool_call in tool_call_buffer
-                    ],
-                    "annotations": [
-                        {
-                            "toolCallId": tool_call["toolCallId"],
-                            "validated": tool_call["validated"],
-                            "isComplete": tool_call["is_complete"],
-                        }
-                        for tool_call in tool_call_buffer
-                    ],
-                }
+                id=uuid.uuid4().hex,
+                role="assistant",
+                createdAt=msg.creation_date,
+                content="",
+                parts=parts,
+                annotations=annotations,
             )
         )
 
@@ -440,4 +444,28 @@ def format_messages_vercel(
         has_more=has_more,
         page_size=page_size,
         results=ordered_messages,
+    )
+
+
+def parse_redis_data(
+    field: str, redis_info: dict[str, tuple[str | None, int]], limit: int
+) -> RateLimitInfo:
+    """From a dictionary containing redis key-value mappings, populate pydantic class."""
+    # Map the field to an actual redis key
+    redis_key = next((key for key in redis_info.keys() if field in key), None)
+
+    # Compute remaining and reset_in
+    remaining = (
+        max(0, limit - int(redis_info[redis_key][0] or 0)) if redis_key else limit
+    )
+    reset_in = (
+        round(redis_info[redis_key][1] / 1000)
+        if redis_key and redis_info[redis_key][1] > 0
+        else None
+    )
+
+    return RateLimitInfo(
+        limit=limit,
+        remaining=remaining,
+        reset_in=reset_in,
     )