+++ frontend/src/components/chat/chat-message-loading.tsx
@@ -4,7 +4,7 @@ import { LoaderPinwheel } from "lucide-react";
 export const ChatMessageLoading = function ChatMessageLoading() {
   return (
     <div className="mt-4 flex justify-start">
-      <Button className="ml-8 mt-1 animate-pulse rounded-full bg-blue-500 p-2.5 hover:bg-blue-500">
+      <Button className="ml-12 mt-1 animate-pulse rounded-full bg-blue-500 p-2.5 hover:bg-blue-500">
         <LoaderPinwheel className="text-black dark:text-white" />
       </Button>
     </div>